Update README.md
Hey Aidan, I made some initial comments on the readme. 
I guess we can strike a balance between what we put in the readme and what we write in the Protocol, so that they complement eachother and do not overlap. I wrote comments for a mix of both. I suggest we make the readme complete enough to run and check the data extracted, and we make the protocol to be more detailed: why did we make these choices? What are the consequences of those choices? How do the files look like (size, variables, unique observations/cultures or paragraphs), etc. 

For the protocol, I also think an initial analysis of what we are covering would be important: number of cultures, number of cultures per source of survival (I don't remember how that variale was called), number of cultures per continent, number of cultures in the EA, distribution of the paragraphs per culture, distribution of the number of paragraphs per subject tab, number of paragraphs in the ethnographic atlas...)